refactor(deploy): split execution.rs god-file into modules
